Add branch support in project analysis

+++ ProjectQualityChecker/Services/IServices/IRepositoryService.cs
@@ -12,7 +12,7 @@ namespace ProjectQualityChecker.Services.IServices
         int Delete(Repository repository);
         int Add(Repository repository);
         Repository Create(RepositoryForm repositoryForm);
-        Repository GetById(int repositoryId);
+        Repository GetById(long repositoryId);
         Task<List<Repository>> GetAllAsync();
         IRepository CloneRepository(string repositoryUrl, string branch = null);
         void DeleteRepositoryDirectory(string path);
